backend/pyproject.toml
[project]
name = "graphtog-backend"
version = "0.1.0"
description = "GraphRAG-based document processing backend with Tree of Graphs reasoning"
requires-python = ">=3.10"
license = {text = "MIT"}
authors = [
    {name = "GraphToG Team", email = "team@graphtog.local"}
]
keywords = ["graphrag", "knowledge-graph", "neo4j", "document-processing", "markdown"]

dependencies = [
    "fastapi>=0.104.0",
    "uvicorn[standard]>=0.24.0",
    "sqlalchemy>=2.0.0",
    "psycopg2-binary>=2.9.0",
    "neo4j>=5.14.0",
    "requests>=2.31.0",
    "google-generativeai>=0.3.0",
    "python-multipart>=0.0.6",
    "python-jose[cryptography]>=3.3.0",
    "passlib[bcrypt]>=1.7.4",
    "python-dotenv>=1.0.0",
    "Pillow>=10.0.0",
    "pydantic>=2.0.0",
    "pydantic[email]>=2.0.0",
    "pydantic-settings>=2.0.0",
    "redis>=5.0.0",
    "pandas>=2.0.0",
    "numpy>=1.24.0",
    "networkx>=3.2",
    "scikit-learn>=1.3.0",
    "tiktoken>=0.5.0",
    "bcrypt>=5.0.0",
    "argon2-cffi>=25.1.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.4.0",
    "pytest-cov>=4.1.0",
    "pytest-asyncio>=0.21.0",
    "pytest-benchmark>=4.0.0",
    "pytest-timeout>=2.1.0",
    "pytest-xdist>=3.3.0",
    "pytest-html>=3.2.0",
    "black>=23.0.0",
    "ruff>=0.1.0",
    "mypy>=1.0.0",
    "httpx>=0.24.0",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["app"]

[tool.black]
line-length = 100
target-version = ["py310"]
include = '\.pyi?$'
extend-exclude = '''
/(
  # directories
  \.eggs
  | \.git
  | \.hg
  | \.mypy_cache
  | \.tox
  | \.venv
  | build
  | dist
)/
'''

[tool.ruff]
line-length = 100
target-version = "py310"
select = [
    "E",   # pycodestyle errors
    "W",   # pycodestyle warnings
    "F",   # Pyflakes
    "I",   # isort
    "C",   # flake8-comprehensions
    "B",   # flake8-bugbear
]
ignore = ["E501"]  # Line too long

[tool.mypy]
python_version = "3.10"
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = false
ignore_missing_imports = true

[dependency-groups]
dev = [
    "httpx>=0.28.1",
    "pytest>=8.4.2",
    "pytest-asyncio>=1.2.0",
    "pytest-benchmark>=5.1.0",
]
